feat: add UI improvement
