+++ src/App.js
@@ -1,7 +1,13 @@
 import Main from './components/Main/Main';
+import Footer from './components/Footer/Footer';
 
 const App = () => {
-  return <Main />;
+  return (
+    <>
+      <Main />
+      <Footer />
+    </>
+  );
 };
 
 export default App;
