+++ src/components/Footer/Footer.css
@@ -1,0 +1,36 @@
+footer {
+  height: 5vh;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+footer a {
+  color: black;
+  text-decoration: none;
+}
+
+footer a img {
+  padding: 5px;
+  height: 30px;
+  cursor: pointer;
+  margin-right: 25px;
+  border-radius: 15px;
+  border: 3px solid black;
+  background-color: white;
+  border-bottom-right-radius: 0;
+}
+
+footer span {
+  height: 30px;
+  cursor: pointer;
+  font-weight: bold;
+  margin-left: 25px;
+  border-radius: 15px;
+  letter-spacing: 1px;
+  font-family: fantasy;
+  padding: 10px 5px 10px 5px;
+  border: 3px solid black;
+  background-color: white;
+  border-bottom-left-radius: 0;
+}
