+++ src/components/Main/Main.css
@@ -1,5 +1,4 @@
 .main {
-  height: 100vh;
-  background: url('../../assets/dark_wall.webp') repeat;
+  height: 90vh;
   font-family: sans-serif;
 }
