+++ src/components/Match/Match.css
@@ -1,5 +1,5 @@
 .match {
-  top: 50%;
+  top: 40%;
   left: 50%;
   position: absolute;
   transform: translate(-50%, -50%);
